fix: reconcile latest main verification gates
