+++ pkg/wire/cli_test.go
@@ -14,12 +14,40 @@ import (
 	platformprocess "github.com/portpowered/infinite-you/pkg/platform/process"
 	serviceedges "github.com/portpowered/infinite-you/pkg/services/edges"
 	factorydefinitions "github.com/portpowered/infinite-you/pkg/services/factory_definitions"
+	"github.com/portpowered/infinite-you/pkg/services/recordings"
 	"github.com/portpowered/infinite-you/pkg/transports/cli/completionprojection"
 	factorycli "github.com/portpowered/infinite-you/pkg/transports/cli/factory"
 	cliobservation "github.com/portpowered/infinite-you/pkg/transports/cli/observation"
 )
 
 type processCommandRunner struct{}
+
+func TestCLIRunDefaultsRetainWireSelectedRecordingTargetPlanner(t *testing.T) {
+	t.Parallel()
+
+	planner := recordings.LiveRecordingTargetPlannerFunc(func(recordings.LiveRecordingTargetRequest) (recordings.LiveRecordingTarget, error) {
+		return recordings.LiveRecordingTarget{}, nil
+	})
+	defaults := provideCLIRunDefaults(planner)
+	if defaults.RecordingTargetPlanner == nil {
+		t.Fatal("CLI run defaults dropped the Wire-selected recording target planner")
+	}
+}
+
+func TestProductionLiveRecordingTargetPlannerIsUsable(t *testing.T) {
+	t.Parallel()
+
+	target, err := provideLiveRecordingTargetPlanner().PlanLiveRecordingTarget(recordings.LiveRecordingTargetRequest{
+		HomeDir:           t.TempDir(),
+		ReportedSessionID: "~default",
+	})
+	if err != nil {
+		t.Fatalf("PlanLiveRecordingTarget: %v", err)
+	}
+	if target.ServicePath == "" || target.ReportedPath == "" || target.ServicePath == target.ReportedPath {
+		t.Fatalf("target = %#v, want distinct runtime template and reported paths", target)
+	}
+}
 
 func (*processCommandRunner) Run(
 	context.Context,
